Initial work on allowing a user to select playback device
Had to move to using DirectSoundOut rather
than WaveOutEvent. By all accounts
DirectSoundOut is a wrappper around a
different windows sound API but offers the
same eventing type of work as WaveOutEvent.
But also allows us to use device GUIDs to
determine what to use for output.

Had to include NAudio in main Blindspot
project now.

Did UI and translations for this feature.